requirements.txt
accelerate
datasets
deepspeed

google-auth
google-auth-oauthlib
huggingface-hub

openai
pandas
rapidfuzz
regex
scipy
sentence-transformers
sentencepiece
sql-metadata
sqlparse
tensorboard

nltk
tqdm
requests
transformers
numpy
scikit-learn
matplotlib
tiktoken
google-generativeai
sqlglot
unidecode
anytree
loguru

aiosqlite
pyserini
faiss-cpu
func_timeout
networkx
ydata-profiling
